[high-level api] remove unused code.
